requirements.txt
asgiref==3.5.0
Django==4.0.3
pytz==2022.1                 # Cross platform timezone calculations
sqlparse==0.4.2              # Parsing, splitting and formatting SQL statements
django-filter==21.1
djangorestframework==3.13.1
Pillow==9.0.1                # Image processing
psycopg2==2.9.3              # PostgreSQL database adapter

